+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
 name = "ddcUtils"
-version = "1.0.13"
+version = "1.0.14"
 description = "Few Utilities"
 readme = "README.md"
 license = "MIT"
